Ajout de la gestion des fichiers utilisateur

+++ backend/controllers/routerBook.js
@@ -1,27 +1,76 @@
 /** Logique métier de chaque route */
+const fs = require('fs');
 const Book = require('../models/Book');
 
 exports.createBook = (req, res, next) => {
-  delete req.body.userId;
+  const bookObject = JSON.parse(req.body.book);
+  // eslint-disable-next-line no-underscore-dangle
+  delete bookObject._id;
+  // eslint-disable-next-line no-underscore-dangle
+  delete bookObject._userId;
   const book = new Book({
-    ...req.body,
+    ...bookObject,
+    userId: req.auth.userId,
+    imageUrl: `${req.protocol}://${req.get('host')}/images/${
+      req.file.filename
+    }`,
   });
   book
     .save()
-    .then(() => res.status(201).json({ message: 'Livre enregistré!' }))
+    .then(() => res.status(201).json({ book }))
     .catch((error) => res.status(400).json({ error }));
 };
 
 exports.modifyBook = (req, res, next) => {
-  Book.updateOne({ _id: req.params.id }, { ...req.body, _id: req.params.id })
-    .then(() => res.status(201).json({ message: 'Livre modifié!' }))
-    .catch((error) => res.status(400).json({ error }));
+  const bookObject = req.file
+    ? {
+        ...JSON.parse(req.body.book),
+        imageUrl: `${req.protocol}://${req.get('host')}/images/${
+          req.file.filename
+        }`,
+      }
+    : { ...req.body };
+  // eslint-disable-next-line no-underscore-dangle
+  delete bookObject._userId;
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      // eslint-disable-next-line eqeqeq
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Non autorisé' });
+      } else {
+        Book.updateOne(
+          { _id: req.params.id },
+          { ...bookObject, _id: req.params.id }
+        )
+          .then(() => res.status(200).json({ message: 'Livre modifié' }))
+          .catch((error) => res.status(401).json({ error }));
+      }
+    })
+    .catch((error) => {
+      res.status(400).json({ error });
+    });
 };
 
 exports.deleteBook = (req, res, next) => {
-  Book.deleteOne({ _id: req.params.id })
-    .then(() => res.status(201).json({ message: 'Livre supprimé!' }))
-    .catch((error) => res.status(400).json({ error }));
+  Book.findOne({ _id: req.params.id })
+    .then((book) => {
+      // eslint-disable-next-line eqeqeq
+      if (book.userId != req.auth.userId) {
+        res.status(401).json({ message: 'Non autorisé' });
+      } else {
+        const filename = book.imageUrl.split('/images/')[1];
+        fs.unlink(`images/${filename}`, () => {
+          Book.deleteOne({ _id: req.params.id })
+            .then(() => {
+              res.status(200).json({ message: 'Livre supprimé' });
+            })
+            .catch((error) => res.status(401).json({ error }));
+        });
+      }
+    })
+    .catch((error) => {
+      res.status(500).json({ error });
+    });
 };
 
 exports.getOneBook = (req, res, next) => {
